Move shortuuid to utils.ts (#523)
